링크 설정 완료, 헤더,푸터 assets로 이동, assets에 함수 모아둘 script파일 작성, 안맞는 ui 소량 수정
