+++ morningstar-library/.settings/org.eclipse.wst.common.component
@@ -1,11 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="Morningstar-Bookstore">
+        
+    <wb-module deploy-name="morningstar-library">
+                
         <wb-resource deploy-path="/" source-path="/WebContent" tag="defaultRootSource"/>
+                
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src"/>
+                
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/WebContent/Control"/>
+                
         <wb-resource deploy-path="/" source-path="/WebContent/templates/user"/>
+                
         <wb-resource deploy-path="/" source-path="/WebContent/templates/admin"/>
-        <property name="java-output-path" value="/Morningstar-Bookstore/build/classes"/>
+                
+        <property name="java-output-path" value="/morningstar-library/build/classes"/>
+                
         <property name="context-root" value="/"/>
+            
     </wb-module>
+    
 </project-modules>
