+++ morningstar-library/WebContent/assets/footer/footer.jsp
@@ -1,0 +1,35 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=1280">
+    <script src="https://cdn.tailwindcss.com"></script>
+    <script src="https://unpkg.com/flowbite@1.5.3/dist/flowbite.js"></script>
+    <title>샛별문고</title>
+</head>
+<body>
+	<div id="footer">
+        <footer class="max-w-[1300px] mx-auto border-t border-slate-900/5 bg-white">
+            <div class="flex flex-row justify-between mx-auto w-full px-10">
+                <div class="float-left mt-7 mb-7">
+                    <p class="text-gray-500 text-sm font-bold mb-2">
+                        한신대학교
+                    </p>
+                    <p class="text-gray-500 text-sm font-normal mb-2">
+                        6조 | 이승은 이학선 전태준 정민구 정규민
+                    </p>
+                    <p class="text-gray-500 text-sm font-normal">
+                        Copyright ⓒ 6조
+                    </p>
+                </div>
+                <div class="content-center justify-end">
+                    <img src="https://www.hs.ac.kr/sites/testPub/images/logo.jpg">
+                </div>
+            </div>
+        </footer>
+    </div>
+</body>
+</html>
