+++ morningstar-library/WebContent/assets/script/admin_script.js
@@ -1,0 +1,5 @@
+/*<script src="../assets/script/admin_script.jsp"></script> */
+
+function delet_user(){
+	
+}
