+++ morningstar-library/WebContent/assets/script/user_script.js
@@ -1,0 +1,1 @@
+/*<script src="../assets/script/user_script.jsp"></script> */
