+++ morningstar-library/src/control/AddCartServlet.java
@@ -9,7 +9,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import model.BookDTO;
+import model.Book;
 import model.Pair;
 
 /**
