+++ morningstar-library/src/model/Cart.java
@@ -6,11 +6,11 @@ import java.util.Vector;
 public class Cart {
 
 	private String userID;
-	private Vector<Pair<BookDTO, Integer, Integer>> books;
+	private Vector<Pair<Book, Integer, Integer>> books;
 	private int totalBookCount;
 	private int totalAmount;
 	
-	public Cart(String userID, Vector<Pair<BookDTO, Integer, Integer>> books, int totalBookCount, int totalAmount) {
+	public Cart(String userID, Vector<Pair<Book, Integer, Integer>> books, int totalBookCount, int totalAmount) {
 		this.userID = userID;
 		this.books = books;
 		this.totalBookCount = totalBookCount;
@@ -28,7 +28,7 @@ public class Cart {
 	
 	public String toString() {
 		Vector<String> bookNames = new Vector<String>();
-		for(Pair<BookDTO, Integer, Integer> book : books) {
+		for(Pair<Book, Integer, Integer> book : books) {
 			bookNames.add(book.getFirst().getBookName());
 		}
 		
@@ -40,35 +40,19 @@ public class Cart {
 				.toString();
 	}
 	
-	
-	
-	/*getter, setter 메소드*/
 	public String getuserID() {
 		return userID;
 	}
-	public void setuserID(String userID) {
-	    this.userID = userID;
-	}
 	
-	public Vector<Pair<BookDTO, Integer, Integer>> getBooks() {
+	public Vector<Pair<Book, Integer, Integer>> getBooks() {
 		return books;
-	}
-	public void setBooks(Vector<Pair<BookDTO, Integer, Integer>> books) {
-	    this.books = books;
 	}
 	
 	public int getTotalBookCount() {
 		return totalBookCount;
 	}
-	public void setTotalBookCount(int totalBookCount) {
-	    this.totalBookCount = totalBookCount;
-	}
 	
 	public int getTotalAmount() {
 		return totalAmount;
 	}
-	public void setTotalAmount(int totalAmount) {
-	    this.totalAmount = totalAmount;
-	}
-	
 }
